Guarantee warp capability for Shyron via Thunder
If Sword of Thunder is used to get to Shyron, guarantee that the player has access to warp boots or teleport before requiring them to get any additional items from that area of the game.  This is important because if they talk to Asina to teleport back to the start, then there may be no viable return route.

Note: Setting the warp point in the temple only works because the temple entrance is not shuffled.  Were it to be shuffled, we'd need to change some things (and pay closer attention to how the temple NPCs spawn, etc).  Possibly we set a warp point for where ever the temple lives, and just guarantee that it's in a town?