Fix broken pit
The problem is that if the lower level is shuffled and the upper is not then we don't look for a valid location.  We instead just save the destination locations and separately shuffle all the pits for all locations.

An issue that comes from this is that the trap behind Sabera drops the player into a level that doesn't directly connect, so we need to account for that, too.