fix disappearing stoned pair
This was caused by a subtle bug where the NPCs were moved to an invalid Y tile (low nibble $f) due to the relative position of the different kind of stair tile, and faulty tile coordinate arithmetic.  The fix was to write up proper tile arithmetic functions and use it when transfering spawns from one metaloc to another.