simplify Flute of Lime logic
Instead of having Asina give the Flute upon a second dialog, we now add an extra NPC in the throne room.  This simplifies the logic because we can now always get the item without being able to cross water, so the queen can now go into the back room before water crossing is available and not lose access to the slot.