Fix rapid-fire paralysis bug
This was introduced by 458c9fb, which changed ContinuousMagicTable (originally at 36092) to be .reloc, using the new .move directive to copy the data from the original table.  Unfortunately, there was a bug in the implementation of .move where it copied data from the wrong location for non-relative offsets.  The fix is to subtract the chunk's .org in that case.  The effect was that it copied all zeros, effectively making _all_ magics continuous - but since every _other_ noncontinuous magic brings up a menu, it's only observable with paralysis.