Fix waterfall cave logic bug
The vanilla game has a flyable tile in the cave tileset that allows bypassing the stoned NPCs by flying left of them if the upstair they blocked was on a left-hand dead end.  The logic figured this out and expected the player to know about it.  This is the same bug that caused players to get stuck inside walls in the crypt.  The patch has been moved from a pre-parse patch in rom.ts to instead be a dedicated named function in deterministic.ts, which should be easier to find in the future.