Fix tower escalator bug
This bug allowed the player to walk down and up an escalator to effectively skip all fighting on the 2nd and 3rd level of the tower.  The underlying problem was that the seamless:down exits were being dropped, so a side effect is that the exit door to walk out of the tower was also broken once the player left the bottom floor (i.e. once you go from $59 to $5a, there's no going back).  We fix it by tweaking how metalocations parse exits, as well as by adding an extra seamless:down exit to the upstair tiles, since these exits are right on the edge.