Improve tower escalator handling
This patch restructures how the escalators are initially opened/closed when entering a new location.  By consolidating some logic, we now allow (1) player can now backtrack all the way out of the tower after forging Crystalis, and (2) all floors are automatically opened when holding Crystalis.  Note that backtracking two floors without Crystalis will still lose progress, since the floor clear flags are still not actually saved anywhere, though it would theoretically be possible to reuse the dialog flag for this if we wanted to add it to the MapData table.