Add spike logic
Large areas of spike (or marsh) will guarantee either flight, rabbit boots, or leather boots/hazmat (unless NoGasMask is set, in which case money and healing/refresh is also allowed).  This should fix seeds where the first sword is behind a ton of spikes in a shuffled oasis cave, but not affect overworld since the river and desert marshes can be skirted around.  We classify a "large area" as one that is fully surrounded by other pain (or impassible) tiles.