Add a bunch of edge:top exits to cave screens
This fixes an issue where we were getting crashes on the vampire level when the arena wasn't up top, since it was placing a 'c' tile instead of an 'n' - so arbitrary hallways were coming up.  More generally, ever since we widened the top exits, it's not unreasonable for the normal-width edges to actually be used as exits.