Fix mezame entrance coordinate bug
Since there's no exit leading to the start entrance (outside of no-bow mode) it was getting pruned.  We now add a dummy circular exit from the door to itself to prevent the pruning, and also need to make sure that entrance 0 is correctly preserved (it was being skipped because it was == 2 tiles, not < 2 tiles from the expected position) since both start-game and wild warp hardcode these entrances.