Fix vampire spawn glitch
When we randomize maps, we pair up exits, arenas, and other POIs (?), and move any nontrivial spawns to have the same relative position to the corresponding location in the new map.  The vanilla vampire position is proximal to both an arena _and_ the map exit, so it's a toss-up which one he gets moved with.  The hacky fix is to only move bosses with arenas, though maybe we could do better later.