Fix Portoa queen logic bug
A rare edge case with Wh can cause a broken seed if the guard-paralysis flow is required to get the queen's first item, but the player doesn't have access to Portoa proper (i.e. the palace door was moved to a different town). We already account for requiring fortune teller access, but the "queen goes away" trigger is in Portoa right outside the door and does not (currently) move with the palace.

The robust solution is to add Flag 020 to tracking and make it a requirement for getting past the back door guard. This makes the logic ensure the player can force the queen to leave and thus paralyze the front door guard to get into the back room. In the future it would _also_ be nice to move the trigger to be in front of whatever door leads to the palace, since it's not really fun to have to go to _three_ different locations (though finding the fortune teller is reasonable, in my opinion).