Add fixEntranceTriggers pass
This pass has a handful of hardcoded triggers that are tied to entrances, but should move with the exit on the other side.  It should be applied in between shuffling entrances and randomizing maps.  It fixes the "queen-fortune teller" dance for shuffled houses to not require _also_ going back to Portoa to force the queen to leave.  It also opens the door for attaching a _different_ cave to the back of the throne room, since we can keep the logic of getting past the guard as the trigger for getting the queen's first item.