Fix kirisa plant trade-in bug
This was a really subtle issue that can rarely show up when a W?t roll takes multiple attempts to complete.  If an earlier roll shuffled trade-ins then we wrote a different item as "must be a girl to use" directly to PRG data.  If a later attempt succeeds _without_ shuffling trade-ins, then the wrong item will still require Change.  It turns out this doesn't actually break the logic (we correctly account for it) but it's unintended and surprising behavior that should be fixed.  In general, we should avoid writing directly to the PRG data before randomization is complete.