Fix monster placement
This fixes an issue that's been around for quite a while.  Some time
ago, we added an unconditional monster location shuffle, which
occurred before the monster identity shuffle.  Even so, the monster
identity shuffle tried really hard to avoid moving momnsters too far,
but it had some hard-coded adjustments for when flyers were turned
into non-flyers, to move them back in bounds.  Unfortunately, when
this offset was added to an already-moved monster, it sometimes put
them out of bounds (a common but mostly benign side effect).  Worse,
an underflow of the Y coordinate can cause the first byte of the
serialized NpcData table entry to wrap around to $FF, which confuses
the program into thinking there are no more spawns in the location,
potentially hiding treasure chests or wall hitboxes (this showed up
rarely in the ball-of-wind room of vampire cave).

Changes:
  * Keep track of whether a location has had any monsters moved.  If
    so, then always use a fresh monsterPlacer during monster identity
    shuffle.
  * Allow placing monsters within the (now, "preferred") clearance of
    another monster (preferring the largest such distance).  This
    significantly decreases cases where monsters fail to place.