separate out optional HUD updates
The EXP refactor (where we count down instead of up) is a required change, since it rearranges RAM a bit.  But it's possible to make only that change without significant disruptions to the classic HUD - in that case, we just remove the "/00000" part after the EXP number.  The HUD update is always on in the current Flagset object, but if shouldUpdateHud() is changed to return false, then the classic HUD now works correctly.