Fix flag allocation error
The old logic for Flag.alloc() was that it would always return $280.  Now it correctly loops to get to the first actually-unused flag.  The issue was discovered because the Cordel water bridge flag is missing on the east seamless partner, and so it ended up using alloc() and getting the same flag as the one we allocate for East Cave 3.  When that exit connects to Cordel East and you fly over to Amazones, you will see a bridge, but it will disappear when you enter and exit Amazones, since you spawn back on the west side.

This fix does _not_ actually sync the bridges on either side of Cordel.  There is no hitbox to form the bridge when you enter from the east, and we don't currently have any way to sync flags like this.  The new behavior just prevents flag sharing - now there's simply no way to form the bridge from that side, and it will never exist.  This does not cause any major problems, and is now just a known issue.