Fix sabera 2 delay bug.
Sabera has a 1 (or maybe 2) second delay before the fight starts, but our patch to allow reverse-direction on Sabera 2 caused the timer to only clock when the player is standing on the 9th row of the map, leading to sometimes very long delays.  The delay is important because otherwise players stabbing Mesia from behind don't have time to jump away.  This fixes the issue by checking whether the screen is already locked, and if so, lifts the specific tile row requirement.