Fix bad reference in Portoa Queen's dialog when shuffling trade-ins
shuffletrades.ts was trying to reference a piece of dialog that no longer exists, after a change (commit c408b95, Oct. 14th 2022) was made to split the Portoa Queen's dialog across two locations, causing an error any time a seed was generated with Wt on. This updates the reference to point to the correct line of dialog.