Restructure build system.
Move to `make` instead of a custom `build.js` script, and switch `npm run watch` to fire up a simple `nodemon` process to re-run `make` as needed (the original `npm run watch` is available as `npm run watch-tsc` for now, though we would do well to clean this up a bit).

Packaged assembly files are now `tar`'d into a single `.tar.br` file externally to esbuild: this allows `make` to avoid rebuilding them when it's not actually necessary.

Reverts the change where we reinterpret the entire PRG through `crystalis.s`: this file is now moved to a separate `vanilla/` directory and is for reference only (it was slowing down the build considerably, and there didn't seem to be an incremental way to start using it).  We now smudge the patch sources as well, so we can start wholesale copying original routines into it (and should probably rearrange the organization a bit).