Add an `overwrite` prop to chunk interface
This is a step toward ensuring we are explicit about what data is written programmatically.  Because programmatically-written chunks will not have access to the same assembler that knows which exact bytes are written, we will instead "pre-declare" it by reserving the bytes in a "no-overwrite" mode, and then run the programmatic assembler in "require-overwrite" mode.