Fix prison key to not automatically set LeafVillagersRescued flag
This flag is set when you get the paralysis check (which requires killing K1).  Setting it prematurely can cause confusing dialog (e.g. Zebu saying "Use your power to open the way" even after the abduction is triggered, or the villagers saying "Thank you for saving me" even when they were never kidnapped in the first place).  Note that this does not actually prevent the latter, but it makes it less bad.  To really fix it, we should probably change the "rescued" trigger to be set by talking to the elder.  But that's a bigger change that I don't want to risk right now.