Fix black holes in sea bug
We had a bug in Metascreens.renumber where we were marking the old screen as unused, even if there were tilesets that weren't included in the renumbering and were still using the old version of the screen.  This didn't cause any problems before, but when we started reserving all the programmatically-written ROM addresses (and zeroing them out), it caused a bug where screen $80 was being left as unused and no longer being initialized correctly to all-$80s (typically blank, but water for the sea tileset).  But since the screen was still in use (we only renumbered caves, not sea), it never actually got written correctly.  By instead checking explicitly for any remaining uses at the end, we can determine whether it's correct to mark the old screen as unused.