Teach smudger how to handle assignments
This allows a new pattern where we define a variable immediately after referencing it (note: there are some potential issues with unknown size for forward-references; we need to be careful that any non-immediate usages are explicitly declared as ZP addresses, either with .importzp or by using the `<` operator).  This provides some self-documentation, but also allows extracting the assignments into an initialization file containing addresses from which to read the data, and then the same name can be later exported for writes.

As an initial PoC, I've applied it to a handful of bosses's music constants.