Refactor shuffleInternal into separate function calls
Create three new functions: preItemFill, createGraph, and postItemFill, and replace most of the contents of shuffleInternal with calls to those three functions.  shuffleInternal is still responsible for the actual item fill, but I could see an argument for breaking that into it's own function as well.

The purpose of this change is to start laying the groundwork for Archipelago integration. The high-level concept is that the Archipelago framework will call into the existing TypeScript code for preItemShuffle and createGraph, then do it's own item fill across the multiworld, and return to the TypeScript for postItemShuffle, which will eventually produce a patch file instead of a patched rom file.