Read predetermined shuffle data from an .apcrys file
Adds function readAPCrysFile to appatch.ts, which takes a filepath to an .apcrys file, then reads from that file a variety of predetermined random selections, a flagstring, and a seed. Additionally adds the --apPatch=FILE command line argument which, when provided with a path to an .apcrys file, will call the new functionality, then proceed with randomization using the flags, seed, and selections provided by the file.