Make MoveObjectWithSpeedAndDirection into .reloc on same $1a page
Note that we had some difficulty because it's easy to "leave behind" a patch from a separate file.  We should go through all the .org-style patches and mark up the vanilla disasm with comments annotating exactly where we've made modifications so that we don't forget to include them in the future.