Initial 1.0 implementation of Archipelago patch support (#154)
* Replace some randomness with predetermined inputs, if present.

* Read predetermined shuffle data from an .apcrys file

Adds function readAPCrysFile to appatch.ts, which takes a filepath to an .apcrys file, then reads from that file a variety of predetermined random selections, a flagstring, and a seed. Additionally adds the --apPatch=FILE command line argument which, when provided with a path to an .apcrys file, will call the new functionality, then proceed with randomization using the flags, seed, and selections provided by the file.

* notes on Archipelago

Added a whole bunch of notes on my approach to Archipelago implementation. Laid out future feature work as well.

Added RAM usage notes as the relate to Archipelago (subject to change until the client is fully implemented).

* ASM changes for Archipelago.

-separates 1xx slot flags from 2xx item flags so that they aren't always updated at the same time.
-adds an hook into the main game loop to read modified RAM data to grant items to the player.
-skips the check-to-item-get conversion that normally happens when getting an item in the randomizer.
-adds a message for finding an APItem.

* Add web interface for Archipelago patching.

* Further separate Slot Flags and Item Flags

This commit modifies PatchGrantItemInRegisterA to check either the slot flags or the item flags based the current state of the Archipelago Item flag (i.e. it checks item flags when receiving an item from the Archipelago client, and checks slot flags when checking a location in the game).

* Add wildwarp ids to the list of things that can be predetermined.

* Tweaks to mimic spawning and chest despawning

Chests now correctly despawn after collecting them without needing to reload the entire room (which previously led to enemies abruptly respawning and teleporting).

Mimics now spawn in either an unused slot or an enemy slot, meaning the worst that happens is an enemy on screen gets replaced with a mimic.  They will also always spawn just above the player, now. Finally, they should no longer glitch the player out when received while airborne.

* Add ShuffleData support for Wg (Shuffle Goa)

* Fix dialog for Archipelago

Refactors fixdialog.ts to separate item wants from item grants, and replaces all the item grant item names with APItem when patching from Archipelago.

Also incidentally fixes the condition for skipping item shuffle when patching from Archipelago.

* Force a flier in Lime Tree Valley when patching from Archipelago.

This guarantees Rage skip is definitely possible, if enabled, which is necessary because Archipelago already assumed it.

* Despawn invisible chests when collecting them in AP.

* Put ARCHIPELAGO on the title screen in place of RANDOMIZER when patching from Archipelago.

* Clear Archipelago flags on cold boot to prevent crashes caused by uninitialized data

* Update Archipelago notes

-Replaced in-progress work plan with end-user process description, now that 1.0 is roughly done.

-Updated to-do list.

-Added musings about how to possibly bake some data into the ROM for local items.

* Accept statue hint information from Archipelago patch data

---------

Co-authored-by: CodeGorilla <3672561+Ars-Ignis@users.noreply.github.com>