clean up realiased labels
There was a hardcoded address in 01_init.s that was simply a different name for MainGameLoopJump_01 (or whatever) and it wasn't correctly updated by the OVERRIDE, causing us to jump into garbage code (in this case it caused an extra frame of slowdown, but it could have easily caused a crash instead).  By eliminating them and consolidating all the labels on a single canonical name, we will be less likely to miss such an update in the future.