Allow for the user to input a Crystalis ROM dumped from the SNK 40th Anniversary Collection available for purchase on Steam.
It's literally one bit of difference, so I added a second allowed CRC32, and then added an entry to the static ROM adjustments that will change the 40th anniversary version into the original NES cart version so that the final checksum is the same.

This should have been a merge from snk-40th, but conflict resolution wanted to push ap-patch into snk-40th, which I'm trying to keep clean for the PR, so instead I'm duplicating the work.