AP Crystalis 2.0 (#159)
* Added new presets for Wha!?, Wham!, and Tournament 2024 (#153)

Add new presets for Wha!?, Wham!, and Tournament 2024.

---------

Co-authored-by: CodeGorilla <3672561+Ars-Ignis@users.noreply.github.com>

* add justchris Trophy

commemorate 2024 tournament winner

* Allow the website to accept the Crystalis ROM from the SNK 40th Anniversary Collection available on Steam (#157)

Allow for the user to input a Crystalis ROM dumped from the SNK 40th Anniversary Collection available for purchase on Steam.

It's literally one bit of difference, so I added a second allowed CRC32, and then added an entry to the static ROM adjustments that will change the 40th anniversary version into the original NES cart version so that the final checksum is the same.

---------

Co-authored-by: CodeGorilla <3672561+Ars-Ignis@users.noreply.github.com>

* Add support for shuffling houses and shuffling areas via predetermined input, to support Wh and Wa in Archipelago.

* Update Archipelago notes/to-do list.

* Move AP communication bytes into save data

There was previously a small chance that BizHawkClient could send an item to the game and update the received item index, then the game would save before actually getting the item in-game, saving the updated index but not the item itself, causing the item to get permanently lost on a reset. Moving the communication bytes into the save data means now in that situation the item will be immediately received upon reloading the save file.

* Default to matching .apcrys file name when patching for AP using CLI

* fix the stupid line endings that I didn't ask for

* Fix broken screen mode when grabbing the Giant Insect chest

* Fix item handling errors introduced by giant insect fix

---------

Co-authored-by: CodeGorilla <3672561+Ars-Ignis@users.noreply.github.com>
Co-authored-by: Stephen Hicks <shicks@users.noreply.github.com>