Fix Zebu's dialog in Zebu's Cave
In a rare case, Wa seeds could break because the logic expects the player to get [Refresh] from Zebu directly after using the Windmill Key, instead of getting it from the invisible trigger in Wind Valley. This specifically caused problems when the player was forced to trigger the Leaf Kidnapping first, as that dialog would take precedence over the dialog to grant the item. The proposed fix is to just reorder the dialog so that the check for having used the Windmill Key always comes first.